backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
networkx==3.4.2
numpy==2.1.3
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==4.25.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
setuptools==80.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
python-multipart